fix:fix invalid firciton

+++ frameworks/core/components_ng/pattern/scrollable/scrollable_pattern.h
@@ -342,7 +342,7 @@ public:
         return scrollSource_;
     }
 
-    int32_t GetCurrentVelocity() const
+    float GetCurrentVelocity() const
     {
         return currentVelocity_;
     }
